fix(dp): address Copilot review on PR #188
Resolves all 7 inline comments from Copilot's re-review of PR #188.

**Code (3):**

- `Poisson.__post_init__` (`opaque-dpsgd/.../accounting/amplification/_poisson.py`):
  validate that `truncated_batch_size` and `dataset_size` are either both
  None or both set, plus per-field bounds. Direct construction and
  deserialization can no longer pass an unpaired pair into
  `_native.truncated_poisson_gaussian_pld` and fail at PLD time. Factory
  drops its now-redundant duplicate validation.

- `fingerprint_per_group_max_norm` (`opaque-dpftrl/.../noise/_distributed.py`):
  return a 64-bit int instead of `int_from_bytes(...) / 2**53`. Float-precision
  loss could let cross-rank `PerGroup` mismatches slip past
  `assert_scalar_equal`; `MFNoiseState._first_max_norm_sync_fingerprint`
  retyped as `int | None`. The DDP sync uses exact int equality.

- `band_mf` / `BandMf.__post_init__` (`opaque-dpftrl/.../mechanisms/_band_mf.py`):
  reject empty `coefficients` tuple. Empty would set `bands == 0` and silently
  zero out downstream `MfPoisson` privacy cost
  (`num_groups = ceil(n_steps / bands)`).

**Docs (4):**

- `docs/api/noise.md`: drop the bogus `opaque.noise_allocation` reference
  (the module has a leading underscore — `opaque._noise_allocation` — and
  is internal). Describe the behaviour without exposing the private path.

- `docs/{api,user-guide}/accounting.md`: refresh both namespace tables and
  the corresponding sub-sections to the current API (`poisson` with
  `truncated_batch_size` / `dataset_size` for DP-SGD, `poisson` with
  `n_steps` for DP-FTRL, `band_mf(..., coefficients=...)`,
  `balls_in_bins(num_bins, n_steps)`, plus `mf_identity`). Removes all
  remaining `cyclic_poisson` / `truncated_poisson` / `num_groups`
  references in these two files.

+++ packages/opaque-dpftrl/src/opaque/dpftrl/noise/_engine.py
@@ -63,7 +63,7 @@ class MFNoiseState(NoiseState):
     _step_counter: int
     _rng_key: RngKey
     _first_max_norm: float | PerGroup | None = None
-    _first_max_norm_sync_fingerprint: float | None = None
+    _first_max_norm_sync_fingerprint: int | None = None
 
 
 def _internal_compute_dtype(dtype: torch.dtype) -> torch.dtype:
